LayoutTests/http/tests/eventloop/idle-callback-not-affected-by-timers-in-cross-origin-expected.txt

This tests requestIdleCallback is not affected by timers scheduled in other origins.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS didRunIdleCallback is true
PASS successfullyParsed is true

TEST COMPLETE
